Added warmup request to start for Config loading.
It seems as if /_ah/start is not being called when
a new instance is started or when an instance is
restarted. I've added warmup requests to the same
StartupHandler function for more added security
that the global config is loaded from GCS after a
new instance is created or reloaded.