Fix null pointer if key from popusers exits cache
This is bound to be buggy because all the keys
could leave memcache before popusers does.
routes.go#62 could cause an infinite loop if all
keys are exhausted.

Just needed something that works in production
because 3am on a school night and I'll fix it
all in the upcoming refactor.