Track Discovered date.
I cannot believe how I forgot to add the Discovered
field to be tracked with each user. On my dev
server, the entities were populated to the
datastore with the current time by default so I
didn't notice the stat wasn't being tracked, but
it makes no sense that my dev server adds the
current time as a default value but the prod
datastore does not.