App modularization refactor.
I've finally gotten around to cleaning up the
code base a bit and making things a little more
readable and extensible. We're not really
following the Go spec to a T yet, but that's the
plan. Aiming for a better grade in Go, so to say.

This is purely a code refactor with no new
features. However, there are five slightly new
behaviors to expect:

* Popular/Featured user(s) are peristed and
cached via memcache and the datastore via
github.com/qedus/nds.

* userCard.html now works with the vine.User type
and not the raw Vine endpoint response. This was
a 1 line change.

* vine.go has been removed and turned into an
independent pkg at github.com/austindizzy/vine-go
so that it may grow independently of Davine.
I look to add all Vine API endpoints to it for use
in other projects and so that it may be of use to
others wanting to use the Vine API in Golang.

* Automatic emails will now contain a configured
name (e.g. Example <abc@example.com> as
opposed to just abc@example.com <abc@example.com>

* config.Load is called during each warmup
request to save the config file in memory for
quicker config lookup times.